requirements.txt
flask
numpy
pillow
torch
torchvision
gunicorn
